requirements.txt
fastapi>=0.103,<1
uvicorn[standard]>=0.23
sqlalchemy>=2,<3
psycopg2-binary>=2.9
python-dotenv>=1
python-telegram-bot==20.7
pillow>=10
pydantic>=2,<3
alembic>=1.12
gunicorn>=21
cloudinary>=1.36
